add AnnotationLayersEncoderDecoder.get_metric(); move AnnotationLayerMetric to taskmodule components and align with interface of torchmetrics.Metric
